requirements.txt
Django==2.1.2
django-debug-toolbar==1.10.1
django-resized==0.3.9
django-reverse-admin==2.1.1
django-tinymce==2.7.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pytz==2018.5
sorl-thumbnail==12.5.0
sqlparse==0.2.4
